# frontend/public/assets/swagger.json
{"swagger":"2.0","info":{"title":"Wikonnect community docs","version":"0.4.0","description":"Wikonnect community docs"},"paths":{"/api/v1/users/{id}":{"get":{"tags":["Authentication"],"summary":"list a single user on the platform","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"id","in":"path","required":true,"type":"string","description":"The users id"}],"responses":{"200":{"description":"successful operation","schema":{"type":"String","items":{"$ref":"#/definitions/GetAUser"}}}}}},"/users":{"get":{"tags":["Authentication"],"summary":"list all user on the platform","consumes":["application/json"],"produces":["application/json"],"parameters":[]}},"/users/{id}/profile-image":{"post":{"tags":["Authentication"],"summary":"upload user profile pic","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","description":"upload user profile pic","schema":{"$ref":"#/definitions/null"}}]}},"/api/v1/users":{"post":{"tags":["Authentication"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","required":true,"schema":{"$ref":"#/definitions/PostAUser"}}]}},"/api/v1/auth":{"post":{"tags":["Authentication"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","required":true,"schema":{"$ref":"#/definitions/PostLoginAUser"}}]}},"/users/{id}":{"put":{"tags":["Authentication"],"summary":"edit users data on the platform","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","description":"edit users data on the platform","schema":{"$ref":"#/definitions/PutAUser"}}],"responses":{"200":{"description":"successful operation","schema":{"type":"String","items":{"$ref":"#/definitions/PutAUser"}}}}}},"/{comment_id}":{"get":{"tags":["ChapterComments"],"consumes":["application/json"],"produces":["application/json"],"parameters":[]},"put":{"tags":["ChapterComments"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","schema":{"$ref":"#/definitions/null"}}]}},"/":{"get":{"tags":["ChapterRatings"],"consumes":["application/json"],"produces":["application/json"],"parameters":[]},"post":{"tags":["ChapterRatings"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","schema":{"$ref":"#/definitions/null"}}]}},"/{rating_id}":{"delete":{"tags":["ChapterRatings"],"consumes":["application/json"],"produces":["application/json"],"parameters":[]},"get":{"tags":["ChapterRatings"],"consumes":["application/json"],"produces":["application/json"],"parameters":[]},"put":{"tags":["ChapterRatings"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","schema":{"$ref":"#/definitions/null"}}]}},"/api/v1/chapters/{id}":{"get":{"tags":["Chapters"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"id","in":"path","required":true,"type":"string","description":"Chapter unique ID"}],"responses":{"200":{"description":"successful operation","schema":{"type":"Object[]","items":{"$ref":"#/definitions/GetAChapter"}}}}},"put":{"tags":["Chapters"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"id","in":"path","required":true,"type":"string","description":"Id - Unique."},{"in":"body","name":"body","required":true,"schema":{"$ref":"#/definitions/PutAChapter"}}],"responses":{"200":{"description":"successful operation","schema":{"type":"Object[]","items":{"$ref":"#/definitions/PutAChapter"}}}}}},"/api/v1/chapters/":{"get":{"tags":["Chapters"],"summary":"Get all chapter and filter using multiple query params","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"id","in":"path","required":false,"type":"string","description":"Optional id"},{"name":"name","in":"path","required":false,"type":"string","description":"Optional name"},{"name":"status","in":"path","required":false,"type":"string","description":"Optional chapter status - published | draft"},{"name":"creatorId","in":"path","required":false,"type":"string","description":"Optional author of a chapter"},{"name":"approved","in":"path","required":false,"type":"boolean","description":"Optional boolean with default being false"}],"responses":{"200":{"description":"successful operation","schema":{"type":"Object[]","items":{"$ref":"#/definitions/GetChapters"}}}}}},"/api/v1/chapters":{"post":{"tags":["Chapters"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","required":true,"schema":{"$ref":"#/definitions/PostAChapter"}}]}},"/api/v1/chapters/{id}/upload":{"post":{"tags":["Chapters"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","schema":{"$ref":"#/definitions/null"}}],"responses":{"200":{"description":"successful operation","schema":{"type":"Object[]","items":{"$ref":"#/definitions/PostAH5PChapter"}}}}}},"/api/v1/chapters/{id}/chapter-image":{"post":{"tags":["Chapters"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","schema":{"$ref":"#/definitions/null"}}],"responses":{"200":{"description":"successful operation","schema":{"type":"Object[]","items":{"$ref":"#/definitions/PostBannerImage"}}}}}},"/courses/{id}":{"delete":{"tags":["Courses"],"consumes":["application/json"],"produces":["application/json"],"parameters":[],"responses":{"200":{"description":"successful operation","schema":{"type":"String","items":{"$ref":"#/definitions/DeleteACourse"}}}}},"get":{"tags":["Courses"],"consumes":["application/json"],"produces":["application/json"],"parameters":[]},"put":{"tags":["Courses"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","schema":{"$ref":"#/definitions/PutACourse"}}],"responses":{"200":{"description":"successful operation","schema":{"type":"String","items":{"$ref":"#/definitions/PutACourse"}}}}}},"/courses/":{"get":{"tags":["Learning_Path"],"consumes":["application/json"],"produces":["application/json"],"parameters":[]}},"/courses":{"post":{"tags":["Courses"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","schema":{"$ref":"#/definitions/PostACourse"}}]}},"/paths":{"post":{"tags":["Courses"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","schema":{"$ref":"#/definitions/PostAPath"}}]}},"/paths/{id}":{"delete":{"tags":["Learning_Path"],"consumes":["application/json"],"produces":["application/json"],"parameters":[],"responses":{"200":{"description":"successful operation","schema":{"type":"String","items":{"$ref":"#/definitions/DeleteLearningPath"}}}}},"get":{"tags":["Learning_Path"],"consumes":["application/json"],"produces":["application/json"],"parameters":[]},"put":{"tags":["Learning_Path"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","schema":{"$ref":"#/definitions/PutAPath"}}],"responses":{"200":{"description":"successful operation","schema":{"type":"String","items":{"$ref":"#/definitions/PutAPath"}}}}}},"/lessons/{id}":{"delete":{"tags":["Lessons"],"consumes":["application/json"],"produces":["application/json"],"parameters":[],"responses":{"200":{"description":"successful operation","schema":{"type":"String","items":{"$ref":"#/definitions/DeleteALesson"}}}}},"get":{"tags":["Lessons"],"consumes":["application/json"],"produces":["application/json"],"parameters":[]},"put":{"tags":["Lessons"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","required":true,"schema":{"$ref":"#/definitions/PutALesson"}}],"responses":{"200":{"description":"successful operation","schema":{"type":"String","items":{"$ref":"#/definitions/PutALesson"}}}}}},"/lessons/":{"get":{"tags":["Lessons"],"consumes":["application/json"],"produces":["application/json"],"parameters":[]}},"/lessons":{"post":{"tags":["Lessons"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","required":true,"schema":{"$ref":"#/definitions/PostALesson"}}]}},"/modules/{id}":{"delete":{"tags":["Modules"],"consumes":["application/json"],"produces":["application/json"],"parameters":[],"responses":{"200":{"description":"successful operation","schema":{"type":"String","items":{"$ref":"#/definitions/DeleteAModule"}}}}},"get":{"tags":["Modules"],"consumes":["application/json"],"produces":["application/json"],"parameters":[]}},"/modules/":{"get":{"tags":["Modules"],"consumes":["application/json"],"produces":["application/json"],"parameters":[]}},"/modules":{"post":{"tags":["Modules"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","required":true,"schema":{"$ref":"#/definitions/PostAModule"}}]}},"api/v1/modules/{id}":{"put":{"tags":["Modules"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","required":true,"schema":{"$ref":"#/definitions/PutAModule"}}]}},"/api/v1/reactions/{id}":{"delete":{"tags":["Reactions"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"id","in":"path","required":true,"type":"string","description":"Reaction Id"}],"responses":{"200":{"description":"successful operation","schema":{"type":"Object[]","items":{"$ref":"#/definitions/DeleteARection"}}}}},"get":{"tags":["Reactions"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"id","in":"path","required":true,"type":"string","description":"Reaction Id"}],"responses":{"200":{"description":"successful operation","schema":{"type":"Object[]","items":{"$ref":"#/definitions/GetAReAction"}}}}},"put":{"tags":["Reactions"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"id","in":"path","required":true,"type":"string","description":"Reaction Id"},{"in":"body","name":"body","required":true,"schema":{"$ref":"#/definitions/PutAReAction"}}],"responses":{"200":{"description":"successful operation","schema":{"type":"Object[]","items":{"$ref":"#/definitions/PutAReAction"}}}}}},"/api/v1/reactions/":{"get":{"tags":["Reactions"],"consumes":["application/json"],"produces":["application/json"],"parameters":[],"responses":{"200":{"description":"successful operation","schema":{"type":"Object[]","items":{"$ref":"#/definitions/GetReactions"}}}}}},"/api/v1/reaction/":{"post":{"tags":["Reactions"],"consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","schema":{"$ref":"#/definitions/PostAReaction"}}]}},"/search/chapter?":{"get":{"tags":["Search"],"summary":"GET result search query using chapter name, description or tags /search?q={query-string-goes-here} Using QUERY string. /search?tags=highschool  Using TAGS.","consumes":["application/json"],"produces":["application/json"],"parameters":[]}}},"definitions":{"GetAUser":{"properties":{"id":{"type":"string","description":"Unique user id"}},"required":["id"]},"PostAUser":{"properties":{"user[username]":{"type":"string","description":"username"},"user[email]":{"type":"string","description":"Unique email"},"user[password]":{"type":"string","description":"validated password"},"user[invitedBy]":{"type":"string","description":"optional auto filled on the form"}},"required":["user[username]","user[email]","user[password]","user[invitedBy]"]},"PostLoginAUser":{"properties":{"username":{"type":"string","description":"username"},"password":{"type":"string","description":"validated password"}},"required":["username","password"]},"PutAUser":{"properties":{"user[object]":{"type":"string","description":"Object data"}},"required":["user[object]"]},"GetAChapter":{"properties":{"id":{"type":"string","description":"Chapter unique ID"},"chapter":{"type":"array","description":"list","items":{"type":"Object"}},"Chapter[object]":{"type":"array","description":"Object data","items":{"type":"Object"}}},"required":["id","chapter","Chapter[object]"]},"chapter":{"properties":{"id":{"type":"string","description":"Id of the chapter"},"name":{"type":"string","description":"Name of the chapter"},"description":{"type":"string","description":"Description of the chapter"},"status":{"type":"string","description":"Status of the chapter (published | draft)"},"approved":{"type":"boolean","description":"boolean with default being false"},"tags":{"type":"object","description":"tags list","$ref":"#/definitions/chapter.tags"}},"required":["id","name","description","status","approved","tags"]},"PutAChapter":{"properties":{"id":{"type":"string","description":"Id - Unique."},"chapter[name]":{"type":"string","description":"Name - Unique."},"chapter[description]":{"type":"string","description":"Description."},"chapter[status]":{"type":"string","description":"modules status - published | draft"},"chapter[approved]":{"type":"boolean","description":"defaults is false"},"chapter[tags]":{"type":"array","description":"Tags list.","items":{"type":"Object"}},"chapter[object]":{"type":"array","description":"Object data","items":{"type":"Object"}}},"required":["id","chapter[name]","chapter[description]","chapter[status]","chapter[approved]","chapter[tags]","chapter[object]"]},"GetChapters":{"properties":{"id":{"type":"string","description":"Optional id"},"name":{"type":"string","description":"Optional name"},"status":{"type":"string","description":"Optional chapter status - published | draft"},"creatorId":{"type":"string","description":"Optional author of a chapter"},"approved":{"type":"boolean","description":"Optional boolean with default being false"},"chapter":{"type":"array","description":"List of chapters","items":{"type":"Object"}}},"required":["chapter"]},"PostAChapter":{"properties":{"chapter[name]":{"type":"string","description":"Name - Unique."},"chapter[description]":{"type":"string","description":"Description."},"chapter[status]":{"type":"string","description":"modules status - options[published | draft]"},"chapter[approved]":{"type":"boolean","description":"defaults is false"},"chapter[tags]":{"type":"array","description":"Tags list.","items":{"type":"Object"}}},"required":["chapter[name]","chapter[description]","chapter[status]","chapter[approved]","chapter[tags]"]},"PostAH5PChapter":{"properties":{"chapter[object]":{"type":"array","description":"Object data","items":{"type":"Object"}}},"required":["chapter[object]"]},"PostBannerImage":{"properties":{"chapter[object]":{"type":"array","description":"Object data","items":{"type":"Object"}}},"required":["chapter[object]"]},"DeleteACourse":{"properties":{"course[object]":{"type":"string","description":"Object data"}},"required":["course[object]"]},"PutACourse":{"properties":{"course[object]":{"type":"string","description":"Object data"}},"required":["course[object]"]},"DeleteLearningPath":{"properties":{"learningpath[object]":{"type":"string","description":"Object data"}},"required":["learningpath[object]"]},"PutAPath":{"properties":{"course[object]":{"type":"string","description":"Object data"}},"required":["course[object]"]},"DeleteALesson":{"properties":{"lesson[object]":{"type":"string","description":"Object data"}},"required":["lesson[object]"]},"PutALesson":{"properties":{"lesson[name]":{"type":"string","description":"Optional Name Unique."},"lesson[slug]":{"type":"string","description":"Optional Slug is Unique and autogenerated."},"lesson[description]":{"type":"string","description":"Optional Description."},"lesson[status]":{"type":"string","description":"lesson status[published or draft]"},"lesson[object]":{"type":"string","description":"Object data"}},"required":["lesson[name]","lesson[slug]","lesson[description]","lesson[status]","lesson[object]"]},"PostALesson":{"properties":{"lesson[name]":{"type":"string","description":"Name - Unique."},"lesson[slug]":{"type":"string","description":"Slug - Unique and autogenerated."},"lesson[description]":{"type":"string","description":"Description."},"lesson[status]":{"type":"string","description":"modules status - published | draft ."},"lesson[creatorId]":{"type":"string","description":"Id of the User."}},"required":["lesson[name]","lesson[slug]","lesson[description]","lesson[status]","lesson[creatorId]"]},"DeleteAModule":{"properties":{"module[object]":{"type":"string","description":"Object data"}},"required":["module[object]"]},"PostAModule":{"properties":{"module[name]":{"type":"string","description":"Name - Unique."},"module[slug]":{"type":"string","description":"Slug - Unique and autogenerated."},"module[description]":{"type":"string","description":"Description."},"module[status]":{"type":"string","description":"modules status - published | draft ."},"module[creatorId]":{"type":"string","description":"Id of the User."}},"required":["module[name]","module[slug]","module[description]","module[status]","module[creatorId]"]},"PutAModule":{"properties":{"module":{"type":"string","description":"List of modules"},"module[slug]":{"type":"string","description":"Optional Slug is Unique and autogenerated."},"module[description]":{"type":"string","description":"Optional Description."},"module[status]":{"type":"string","description":"modules status[published or draft]"}},"required":["module"]},"module":{"properties":{"name":{"type":"string","description":"Optional Name Unique."}},"required":["name"]},"DeleteARection":{"properties":{"id":{"type":"string","description":"Reaction Id"},"reaction":{"type":"array","description":"list","items":{"type":"Object"}}},"required":["id","reaction"]},"reaction":{"properties":{"id":{"type":"string","description":"Reaction id"},"chapterId":{"type":"string","description":"Chapter id."},"userId":{"type":"string","description":"Authenticated user id"}},"required":["id","chapterId","userId"]},"GetAReAction":{"properties":{"id":{"type":"string","description":"Reaction Id"},"reaction":{"type":"array","description":"list","items":{"type":"Object"}},"reaction[object]":{"type":"array","description":"Object data","items":{"type":"Object"}}},"required":["id","reaction","reaction[object]"]},"PutAReAction":{"properties":{"id":{"type":"string","description":"Reaction Id"},"reaction":{"type":"array","description":"list","items":{"type":"Object"}}},"required":["id","reaction"]},"GetReactions":{"properties":{"reaction":{"type":"array","description":"list","items":{"type":"Object"}},"Reaction[object]":{"type":"array","description":"Reaction object data","items":{"type":"Object"}}},"required":["reaction","Reaction[object]"]}}}